Model, script and animation updates
- Implemented the free asset Allosaurus models and animations into the
project.
- Decoupled the prey and predators from a single script to two different
scripts.
- Fixed small bugs with selection/deselection where text from a
deselected agent would not disappear.
- Added initial simulation model variables including health and
endurance for the prey and predator. The prey currently is the only
agent that has its endurance tied to its maximum movement speed. The
speed is multiplied by the endurance factor which ranges from 0 and 1.
- Endurance is reduced by .01 every second the prey is running. The
endurance is regenerated by .01 every second the prey is moving slower
than the max walk speed.
- KNOWN ISSUE: Prey still runs into corners without turning.