Updated prey, predator, simulation space, camera
Prey:
- Attempted to couple boids modeling with the navmeshagent.
- Prey will attempt to run away from the avg position of predators
chasing it. If not it will check to see if there are any fleeing prey
behind it and flock with them. If there are no fleeing prey it will
attempt to flock with nearby prey. If there are none, the prey will
stop.
- Added TrailRenderer to the Prey prefab. Prey will now leave a red
line/trail behind them indicating their previous path.

Predator:
- Predators will attempt to chase the closest prey.
- WIP: Attempting to create some sort of boids based model for
predators.

Simulation space:
- Doubled the area of the simulation "play area".

Camera:
- Increased the maximum and minimum y-angle of the camera from 45/-45 to
15/-15.