Updates to prey, predator, camera
General:
- Increased the number of Prey and Predators.
- Updated starting positions so Prey are closer together.
- Added new material for Predator trails.
- Added a screenshot key which will save a single screenshot at a time
to the Assets/Screenshots/ directory. This will not work properly if the
simulation is fully built and will need to be updated in the future.

Prey:
- Modified boids values to create more realistic herding behavior.
- Added the ability for Prey to have their endurance and health sapped.
- Added "dead" state for when a Prey no longer has any health.

Predator:
- Updated beahavior, will now attempt to chase Prey directly inside of a
focusRadius. Will attempt attack a Prey inside of a kill radius (which
is smaller than the focusRadius). If no Prey is in a focusRadius it will
attempt to flock with nearby Prey by using boids model.
- Implemented attack animation for when a Prey is inside of the
killRadius.
- Turned on the TrailRenderer so the path of the Predator can now be
seen.